+++ apps/docs/src/content/04-components/form-controls/date-picker/examples/contextualHelp.tsx
@@ -3,7 +3,6 @@ import {
   ContextualHelp,
   ContextualHelpTrigger,
   DatePicker,
-  Heading,
   Text,
   Label,
 } from "@mittwald/flow-react-components";
@@ -11,10 +10,9 @@ import {
 <DatePicker>
   <Label>
     Datum
-    <ContextualHelpTrigger>
+    <ContextualHelpTrigger subject="Datum">
       <Button />
       <ContextualHelp>
-        <Heading>Weitere Informationen</Heading>
         <Text>
           Hier gibt es weitere Informationen, die zu lang
           für die FieldDescription sind.