+++ apps/docs/src/content/04-components/structure/tabs/examples/status.tsx
@@ -1,0 +1,20 @@
+import {
+  AlertIcon,
+  Tab,
+  Tabs,
+  TabTitle,
+} from "@mittwald/flow-react-components";
+
+<Tabs>
+  <Tab>
+    <TabTitle>Mit App verbinden</TabTitle>
+    App Content
+  </Tab>
+  <Tab>
+    <TabTitle>
+      Mit Container verbinden
+      <AlertIcon status="warning" />
+    </TabTitle>
+    Container Content
+  </Tab>
+</Tabs>;