+++ packages/components/src/index/internal.ts
@@ -4,3 +4,5 @@ export * from "@/lib/componentFactory/flowComponent";
 export * from "@/lib/propsContext";
 export * from "@/lib/viewComponentContext";
 export * from "@/lib/types/props";
+export * from "@/status/componentStatus";
+export type * from "@/status/types";